updated unity config and transport to comply with pio core 6.0
