+++ test/test/helpers/run-tests.h
@@ -1,8 +1,17 @@
 #pragma once
 
-#if defined(ARDUINO_AVR_MEGA2560) || defined(ENV_NATIVE)
+#if defined(ENV_NATIVE)
+
 int main(int argc, char **argv) { return tests(); }
+
 #else
+
 void setup() { delay(1000); }
-void loop() { tests(); }
+
+void loop()
+{
+    tests();
+    delay(500);
+}
+
 #endif
