+++ test/test/test_easycomm2_responses/test.cpp
@@ -10,36 +10,6 @@
 #include <unity.h>
 
 
-void test_response_single_line_01()
-{
-    const char *expected_response = "AZ0.0 EL0.0 UP0 upm DN0 dnm";
-    EasycommResponseSingleLine response;
-    easycommResponseSingleLine(&response);
-    memcpy(response.modeUp, "upm", 3);
-    memcpy(response.modeDown, "dnm", 3);
-    char response_str[32];
-    easycommResponseSingleLineSprintf(&response, response_str);
-    TEST_ASSERT_EQUAL_STRING(expected_response, response_str);
-}
-
-
-void test_response_single_line_02()
-{
-    const char *expected_response = "AZ10.0 EL11.0 UP12 x DN13 y";
-    EasycommResponseSingleLine response;
-    easycommResponseSingleLine(&response);
-    response.azimuth = 10;
-    response.elevation = 11;
-    response.uplinkFrequency.as.uint32 = 12;
-    response.downlinkFrequency.as.uint32 = 13;
-    memcpy(response.modeUp, "x", 1);
-    memcpy(response.modeDown, "y", 1);
-    char response_str[32];
-    easycommResponseSingleLineSprintf(&response, response_str);
-    TEST_ASSERT_EQUAL_STRING(expected_response, response_str);
-}
-
-
 void test_response_azimuth_01()
 {
     const char *expected_response = "AZ12.4";
@@ -235,47 +205,9 @@ void test_response_velocity_down_01()
 }
 
 
-void test_response_read_config_01()
-{
-    const char *expected_response = "CR1,#+!$abc";
-    EasycommResponseConfigRegister response;
-    easycommResponseConfigRegister(&response);
-    response.registerNumber = 1;
-    memcpy(response.value.as.str, "#+!$abc", 8);
-    char response_str[32];
-    easycommResponseConfigRegisterSprintf(&response, response_str);
-    TEST_ASSERT_EQUAL_STRING(expected_response, response_str);
-}
-
-
-void test_response_get_status_register_01()
-{
-    const char *expected_response = "GS2";
-    EasycommResponseStatusRegister response;
-    easycommResponseStatusRegister(&response);
-    response.status = EasycommStatusMoving;
-    char response_str[32];
-    easycommResponseStatusRegisterSprintf(&response, response_str);
-    TEST_ASSERT_EQUAL_STRING(expected_response, response_str);
-}
-
-
-void test_response_get_error_register_01()
-{
-    const char *expected_response = "GE2";
-    EasycommResponseErrorRegister response;
-    easycommResponseErrorRegister(&response);
-    response.status = EasycommErrorJam;
-    char response_str[32];
-    easycommResponseErrorRegisterSprintf(&response, response_str);
-    TEST_ASSERT_EQUAL_STRING(expected_response, response_str);
-}
-
 int tests()
 {
     UNITY_BEGIN();
-    RUN_TEST(test_response_single_line_01);
-    RUN_TEST(test_response_single_line_02);
     RUN_TEST(test_response_azimuth_01);
     RUN_TEST(test_response_elevation_01);
     RUN_TEST(test_response_uplink_frequency_01);
@@ -292,9 +224,6 @@ int tests()
     RUN_TEST(test_response_velocity_right_01);
     RUN_TEST(test_response_velocity_up_01);
     RUN_TEST(test_response_velocity_down_01);
-    RUN_TEST(test_response_read_config_01);
-    RUN_TEST(test_response_get_status_register_01);
-    RUN_TEST(test_response_get_error_register_01);
     return UNITY_END();
 }
 
@@ -302,4 +231,4 @@ void setUp() {}
 
 void tearDown() {}
 
-#include "../run-tests.h"
+#include "../helpers/run-tests.h"
