+++ test/test/test_easycomm_buffered_tokenizer/test.cpp
@@ -426,4 +426,4 @@ void setUp() {}
 
 void tearDown() {}
 
-#include "../run-tests.h"
+#include "../helpers/run-tests.h"
