+++ test/test/test_easycomm_command_handler/test.cpp
@@ -135,4 +135,4 @@ void setUp() {}
 
 void tearDown() {}
 
-#include "../run-tests.h"
+#include "../helpers/run-tests.h"
