+++ test/test/test_sscanf_float/test.cpp
@@ -41,4 +41,4 @@ void setUp() {}
 
 void tearDown() {}
 
-#include "../run-tests.h"
+#include "../helpers/run-tests.h"
