+++ test/test/transport_arduino_arch_stm32.c
@@ -1,5 +1,5 @@
 #if defined(ARDUINO_ARCH_STM32)
-#include "arch_stm32_transport.h"
+#include "transport_arduino_arch_stm32.h"
 #include "stm32f4xx_hal.h"
 
 #define USARTx USART2
